Fix broken imports/syntax, wire MarketIdentity end-to-end, move odds to Decimal
The package could not be imported: root __init__.py pulled from a
nonexistent module, odds.py/raw_odds.py referenced MarketIdentity
without importing it, and app.py, odds_repository.py, database.py,
and test_odds_repository.py had syntax errors (duplicated methods,
bad indentation, stray SQL semicolon). Fixed all of these and added
a CI workflow that imports the package and runs pytest so this class
of regression is caught automatically.

Also finished wiring MarketIdentity through the collector, app demo,
and storage schema instead of the previous plain "1X2" string, and
switched odds from float to Decimal throughout (models, validation,
analysis, storage) to avoid rounding errors in surebet margins.

Fixed a latent bug found while repairing app.py: demo event start
times were 2 hours outside the matcher's tolerance window relative
to the sample data, and the sample JSON timestamps were naive while
demo events were UTC-aware, which would raise TypeError on
comparison. Added UTC offsets to the sample data and aligned start
times so the demo actually matches events end-to-end.

EventMatcher now records candidate ids in the ambiguous-event reason
for easier debugging. Removed two empty placeholder docs.

+++ src/anomaly_detection_engine/models/raw_odds.py
@@ -1,5 +1,8 @@
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
+
+from anomaly_detection_engine.models.market import MarketIdentity
 
 
 @dataclass(frozen=True)
@@ -12,5 +15,5 @@ class RawEventOdds:
     start_time: datetime
     observed_at: datetime
     market: MarketIdentity
-    odds: dict[str, float]
+    odds: dict[str, Decimal]
     source_timestamp: datetime | None = None
